@@ -1,4 +1,4 @@
-import { Button, Step, StepLabel, Stepper } from "@material-ui/core";
+import { Step, StepLabel, Stepper } from "@mui/material";
 import React from "react";
 
 const fases = [
