@@ -1,6 +1,6 @@
 import React from "react";
 
-import TextField from "@material-ui/core/TextField";
+import { TextField } from "@mui/material";
 import { useSelector } from "react-redux";
 
 import { useTeam, useSocket } from "..";
