@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "@material-ui/core/Button";
+import { Button } from "@mui/material";
 import { useHistory, useLocation } from "react-router-dom";
 
 import API from "../../../api";
