@@ -1,7 +1,7 @@
 import React from "react";
 import { useHistory } from "react-router-dom";
 
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 
 import { RiddleRoutes } from "..";
 
